+++ src/win/camera.rs
@@ -4,52 +4,29 @@ use na::Rotation3;
 use sdl2::keyboard::Keycode;
 use serde::{Deserialize, Serialize};
 
-pub const DEFAULT_CAMERA_UP: Vec3 = Vec3::new(0.0, 0.0, 1.0);
-pub const DEFAULT_CAMERA_DIRECTION: Vec3 = Vec3::new(-1.0, 0.0, 0.0);
-pub const DEFAULT_CAMERA_POSITION: Vec3 = Vec3::new(5.0, 0.0, 0.0);
-pub const DEFAULT_CAMERA_ANCHOR: Vec3 = Vec3::new(0.0, 0.0, 0.0);
-pub const DEFAULT_NEAR_FACTOR: Float = 1e-4;
-pub const DEFAULT_FAR_FACTOR: Float = 2.0;
-pub const DEFAULT_FOVY: Float = 40.0 * RPD;
-pub const DEFAULT_SENSITIVITY: Float = 1.0;
-pub const SENSITIVITY_CORRECTION: Float = 1e-4;
-pub const DEFAULT_SPEED: Float = 1.0;
-pub const SPEED_CORRECTION: Float = 1.0;
-pub const KEYBOARD_SENSITIVITY_CORRECTION: Float = 1e-2;
+pub const UP: Vec3 = Vec3::new(0.0, 0.0, 1.0);
+pub const DIRECTION: Vec3 = Vec3::new(-1.0, 0.0, 0.0);
+pub const POSITION: Vec3 = Vec3::new(5.0, 0.0, 0.0);
+pub const ANCHOR: Vec3 = Vec3::new(0.0, 0.0, 0.0);
+pub const NEAR_FACTOR: Float = 1e-5;
+pub const FAR_FACTOR: Float = 2.0;
+pub const FOVY: Float = 40.0 * RPD;
+pub const SENSITIVITY: Float = 1.0;
+pub const SENSITIVITY_CORRECTION: Float = 1e-3;
+pub const FREE_SENSITIVITY_CORRECTION: Float = 1e-1;
+pub const FREE_KEYBOARD_SENSITIVITY_CORRECTION: Float = 1e-3;
+pub const SPEED: Float = 1.0;
+pub const SPEED_FAST_FACTOR: Float = 10.0;
+
+pub const KEY_FORWARD: Keycode = Keycode::W;
+pub const KEY_LEFT: Keycode = Keycode::A;
+pub const KEY_BACKWARD: Keycode = Keycode::S;
+pub const KEY_RIGHT: Keycode = Keycode::D;
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum MovementMode {
-    Rotate,
-    Strafe,
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum Direction {
-    Up,
-    Left,
-    Down,
-    Right,
-}
-
-impl Direction {
-    pub fn from_keycode(keycode: Keycode) -> Self {
-        match keycode {
-            Keycode::Up => Self::Up,
-            Keycode::Left => Self::Left,
-            Keycode::Down => Self::Down,
-            Keycode::Right => Self::Right,
-            _ => panic!("Not a valid keycode for direction: {:?}.", keycode),
-        }
-    }
-
-    pub fn to_xy(&self) -> (i32, i32) {
-        match self {
-            Self::Up => (0, 1),
-            Self::Left => (-1, 0),
-            Self::Down => (0, -1),
-            Self::Right => (1, 0),
-        }
-    }
+    Lock,
+    Free,
 }
 
 #[derive(Clone, Debug, Copy, Serialize, Deserialize, PartialEq)]
@@ -63,15 +40,15 @@ pub enum ProjectionMode {
 
 impl Default for ProjectionMode {
     fn default() -> Self {
-        Self::Perspective(DEFAULT_FOVY)
+        Self::Perspective(FOVY)
     }
 }
 
 impl ProjectionMode {
     pub fn matrix(&self, distance: Float, aspect: Float) -> Mat4 {
         let side = distance;
-        let zfar = distance * DEFAULT_FAR_FACTOR;
-        let znear = zfar * DEFAULT_NEAR_FACTOR;
+        let zfar = distance * FAR_FACTOR;
+        let znear = zfar * NEAR_FACTOR;
 
         match self {
             ProjectionMode::Orthographic => {
@@ -96,13 +73,13 @@ pub struct Camera {
 impl Camera {
     pub fn new() -> Self {
         Self {
-            up: DEFAULT_CAMERA_UP,
-            direction: DEFAULT_CAMERA_DIRECTION,
-            position: DEFAULT_CAMERA_POSITION,
-            anchor: DEFAULT_CAMERA_ANCHOR,
+            up: UP,
+            direction: DIRECTION,
+            position: POSITION,
+            anchor: ANCHOR,
             projection: ProjectionMode::default(),
-            movement_mode: MovementMode::Rotate,
-            up_world: DEFAULT_CAMERA_UP,
+            movement_mode: MovementMode::Lock,
+            up_world: UP,
         }
     }
 
@@ -163,46 +140,19 @@ impl Camera {
         self.fix_up();
     }
 
-    pub fn move_command(&mut self, keycode: Keycode, delta_time: Float) {
-        let direction = Direction::from_keycode(keycode);
-        match self.movement_mode {
-            MovementMode::Rotate => {
-                let (x, y) = direction.to_xy();
-                self.rotate_around_anchor(x, y, delta_time)
-            }
-            MovementMode::Strafe => self.strafe(direction, delta_time),
-        };
-    }
-
     pub fn toggle_movement_mode(&mut self) -> MovementMode {
         self.movement_mode = match self.movement_mode {
-            MovementMode::Rotate => MovementMode::Strafe,
-            MovementMode::Strafe => MovementMode::Rotate,
+            MovementMode::Lock => MovementMode::Free,
+            MovementMode::Free => MovementMode::Lock,
         };
         self.movement_mode
     }
 
-    /*
-    pub fn change_speed(&mut self, direction: Direction, delta_time: Float) {
-        let delta_speed = match direction {
-            Direction::Up => 1.0,
-            Direction::Down => -1.0,
-            Direction::Left | Direction::Right => {
-                unreachable!("Camera::change_speed match unreachable.")
-            }
-        } * DEFAULT_SENSITIVITY
-            * DEFAULT_SENSITIVITY_CORRECTION
-            * delta_time as Float;
+    pub fn lock_rotate(&mut self, x: Float, y: Float) {
+        let correction = SENSITIVITY_CORRECTION;
 
-        // self.speed += delta_speed;
-    }
-    */
-
-    pub fn rotate_around_anchor(&mut self, x: i32, y: i32, delta_time: Float) {
-        let sensitivity = DEFAULT_SENSITIVITY * KEYBOARD_SENSITIVITY_CORRECTION * delta_time;
-
-        let m1 = Rotation3::new(self.up_world * x as Float * sensitivity);
-        let m2 = Rotation3::new(-self.right() * y as Float * sensitivity);
+        let m1 = Rotation3::new(self.up_world * x * correction);
+        let m2 = Rotation3::new(self.right() * y * correction);
         let m = m1 * m2;
 
         self.position = self.anchor + m * (self.position - self.anchor);
@@ -211,28 +161,19 @@ impl Camera {
         self.target_anchor();
     }
 
-    pub fn strafe(&mut self, direction: Direction, delta_time: Float) {
-        let sensitivity =
-            DEFAULT_SPEED * KEYBOARD_SENSITIVITY_CORRECTION * SPEED_CORRECTION * delta_time;
-
-        let delta_position = (match direction {
-            Direction::Up => self.front(),
-            Direction::Left => self.left(),
-            Direction::Down => self.back(),
-            Direction::Right => self.right(),
-        }) * sensitivity;
-
-        self.position += delta_position;
+    pub fn free_movement(&mut self, x: Float, y: Float) {
+        let correction = FREE_KEYBOARD_SENSITIVITY_CORRECTION;
+        self.position += (y * self.front() + x * self.right()) * correction;
     }
 
-    pub fn rotate_strafe(&mut self, x: i32, y: i32, delta_time: Float) {
-        let sensitivity = DEFAULT_SENSITIVITY * SENSITIVITY_CORRECTION * delta_time;
-        let mut sphericals = cartesian_to_spherical(&self.direction);
+    pub fn free_rotate(&mut self, x: Float, y: Float) {
+        let correction = SENSITIVITY_CORRECTION * FREE_SENSITIVITY_CORRECTION;
 
-        sphericals[1] -= x as Float * sensitivity;
-        sphericals[2] = (sphericals[2] - y as Float * sensitivity).clamp(-PI / 2.0, PI / 2.0);
+        let m1 = Rotation3::new(self.up * x * correction);
+        let m2 = Rotation3::new(self.right() * y * correction);
+        let m = m1 * m2;
 
-        self.direction = spherical_to_cartesian(&sphericals);
-        self.fix_up();
+        self.up = m * self.up;
+        self.direction = m * self.direction;
     }
 }